SpaceGroup class and new Dialog
The SpaceGroups class is formally independent of Spglib but uses it's
struct to store cyrstal system, bravais lattic, symbol and other
information. The unitcell class now only stores the Hall index
and futher look-up is performed in SpaceGroups

The spacegroupdialog box sets it values from SpaceGroups and will
be used to query and assign the space group of the molecule.

I need to shrink the width of the columns.

Change-Id: Ife5be28233aea4c4b2f26e740d595e63b273c4f0